[整理] Bookmark情報登録のコードを適切な箇所に整理
